+++ src/server/api/routers/judge0/judge0.input.ts
@@ -1,5 +1,11 @@
 import { z } from "zod";
 
+export const judge0SubmissionSchema = z.object({
+  source_code: z.string().min(1, "Source code cannot be empty"),
+  language_id: z.number().positive("Invalid language ID"),
+  stdin: z.string().optional(),
+});
+
 export const judge0ResponseSchema = z.object({
   token: z.string(),
   stdout: z.string().nullable(),
@@ -20,3 +26,4 @@ export const judge0LanguageSchema = z.object({
 
 export type judge0ResponseSchema = z.infer<typeof judge0ResponseSchema>;
 export type judge0LanguageSchema = z.infer<typeof judge0LanguageSchema>;
+export type judge0SubmissionSchema = z.infer<typeof judge0SubmissionSchema>;
