+++ src/server/api/routers/room/room.service.ts
@@ -1,6 +1,6 @@
 import { env } from "@/env";
 import { sendMeetingEmail } from "@/lib/mail/mailer";
-import { generateRoomName } from "@/lib/room-name-generator";
+import { generateName } from "@/lib/generic-name-generator";
 import { db } from "@/server/db";
 
 import type { CreateRoomInput } from "./room.input";
@@ -14,10 +14,11 @@ export async function createRoom(input: CreateRoomInput) {
 
   if (input.hostEmail) {
     const roomLink = `http://localhost:3000/room/${room.id}`;
-    
+
     await sendMeetingEmail({
       roomLink,
-      scheduledDateAndTime: input.startDateTime?.toISOString() ?? new Date().toISOString(),
+      scheduledDateAndTime:
+        input.startDateTime?.toISOString() ?? new Date().toISOString(),
       to: input.hostEmail,
       from: env.SMTP_FROM_EMAIL ?? "noreply@meerkat.com",
       subject: "Your Meerkat Room is Ready",
@@ -30,7 +31,7 @@ export async function createRoom(input: CreateRoomInput) {
 export async function createInstantRoom() {
   const room = await db.room.create({
     data: {
-      roomName: await generateRoomName(),
+      roomName: await generateName(),
       controlMicrophone: true,
       allowPasting: true,
     },
